git-svn-id: https://plugins.svn.wordpress.org/hyper-cache/trunk@72553 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ plugin.php
@@ -72,70 +72,53 @@ $hyper_options = get_option('hyper');
 $hyper_invalidated = false;
 $hyper_invalidated_post_id = null;
 
+// On activation, we try to create files and directories. If something goes wrong
+// (eg. for wrong permission on file system) the options page will give a
+// warning.
 add_action('activate_hyper-cache/plugin.php', 'hyper_activate');
 function hyper_activate() 
 {
-    if (!file_exists(ABSPATH . 'wp-content/hyper-cache')) 
-    {
-        if (!mkdir(ABSPATH . 'wp-content/hyper-cache', 0766)) return new WP_Error('', 'Cannot create the cache directory');
-        if (!touch(ABSPATH . 'wp-content/hyper-cache/test.dat')) return new WP_Error('', 'Cannot write into the cache directory');
-        $file = time();
-        if (!mkdir(ABSPATH . 'wp-content/' . $file, 0766)) return new WP_Error('', 'Cannot create directories in wp-content');
-        rmdir(ABSPATH . 'wp-content/' . $file);
-    }
-	//hyper_cache_invalidate();
-	
-    // Write the advanced-cache.php (so we grant it's the correct version)
+    mkdir(ABSPATH . 'wp-content/hyper-cache', 0766);
+
     $buffer = file_get_contents(dirname(__FILE__) . '/advanced-cache.php');
-    $file = fopen(ABSPATH . 'wp-content/advanced-cache.php', 'w');
+    $file = @fopen(ABSPATH . 'wp-content/advanced-cache.php', 'w');
     if ($file)
     {
         fwrite($file, $buffer);
         fclose($file);	
     }
-    else
-    {
-        return new WP_Error('', 'Cannot create advanced-cache.php into wp-content directory');    
-    }
 }
-
 
 add_action('deactivate_hyper-cache/plugin.php', 'hyper_deactivate');
 function hyper_deactivate() 
 {
-	if (file_exists(ABSPATH . 'wp-content/advanced-cache.php')) unlink(ABSPATH . 'wp-content/advanced-cache.php');
-	if (file_exists(ABSPATH . 'wp-content/hyper-cache-config.php')) unlink(ABSPATH . 'wp-content/hyper-cache-config.php');
+	unlink(ABSPATH . 'wp-content/advanced-cache.php');
+	unlink(ABSPATH . 'wp-content/hyper-cache-config.php');
 
-	if (is_dir(ABSPATH . 'wp-content/hyper-cache'))
-	{
-		$path = ABSPATH . 'wp-content/' . time();
-		rename(ABSPATH . 'wp-content/hyper-cache', $path);
-
-		hyper_delete_path( $path );
-	}
+    // We can safely delete the hyper-cache directory, is not more used at this time.
+    hyper_delete_path(ABSPATH . 'wp-content/hyper-cache');
 }
 
 add_action('admin_head', 'hyper_admin_head');
-function hyper_admin_head() {
+function hyper_admin_head() 
+{
 	add_options_page('Hyper Cache', 'Hyper Cache', 'manage_options', 'hyper-cache/options.php');
 }
 
-function hyper_cache_invalidate_publish_post() 
-{
-    hyper_log("Called global invalidate for publish_post");
-    hyper_cache_invalidate();
-}
-
-function hyper_cache_invalidate($force=false) 
+// Completely invalidate the cache. The hyper-cache directory is renamed
+// with a random name and re-created to be immediately available to the cache
+// system. Then the renamed directory is removed.
+// If the cache has been already invalidated, the function doesn't anything.
+function hyper_cache_invalidate() 
 {
 	global $hyper_options, $hyper_invalidated;
 	
-    hyper_log("Called global invalidate");
-    if ($hyper_invalidated) hyper_log("Already invalidated");
-    if ($hyper_invalidated) return;
-    
-	//if (!$force && $hyper_options['not_expire_on_actions']) return;
-	if (!$force) return;
+    //hyper_log("Called global invalidate");
+    if ($hyper_invalidated) 
+    {
+        //hyper_log("Already invalidated");
+        return;
+    }
 	
     $hyper_invalidated = true;
 	$path = ABSPATH . 'wp-content/' . time();
@@ -146,7 +129,7 @@ function hyper_cache_invalidate($force=false)
 
 function hyper_cache_invalidate_post($post_id) 
 {
-    hyper_log("Called post invalidate for post id $post_id");
+    //hyper_log("Called post invalidate for post id $post_id");
     hyper_delete_by_post($post_id);
 }
 
@@ -155,14 +138,14 @@ function hyper_cache_invalidate_post_status($new, $old, $post)
     global $hyper_invalidated;
     
     $post_id = $post->ID;
-    hyper_log("Called post status invalidate for post id $post_id from $old to $new");
+    //hyper_log("Called post status invalidate for post id $post_id from $old to $new");
 
     // The post is going online or offline
     if ($new != 'publish' && $old != 'publish') return;
 
-    hyper_log("Start global invalidation");
+    //hyper_log("Start global invalidation");
     
-    if ($hyper_invalidated) hyper_log("Already invalidated");
+    if ($hyper_invalidated) //hyper_log("Already invalidated");
     if ($hyper_invalidated) return;
         
     $hyper_invalidated = true;
@@ -174,7 +157,7 @@ function hyper_cache_invalidate_post_status($new, $old, $post)
 
 function hyper_cache_invalidate_comment($comment_id, $status=1) 
 {
-    hyper_log("Called comment invalidate for comment id $comment_id and status $status");
+    //hyper_log("Called comment invalidate for comment id $comment_id and status $status");
     if ($status != 1) return;
     hyper_delete_by_comment($comment_id);
     //hyper_cache_invalidate();
@@ -187,62 +170,66 @@ function hyper_delete_by_comment($comment_id)
     hyper_delete_by_post($post_id);
 }
 
+
+// Delete files in the cache based only on a post id. Only the post cached
+// page is deleted (with its mobile versions) and the home page of the blog.
+// If the cache has been already invalidated, of the specified post has already
+// been invalidate the function return doing nothing. This behaviour protect
+// from multi invalidation caused by actions fired by WP.
+// The invalidation doesn't take place is the post is not in "publish" status.
 function hyper_delete_by_post($post_id)
 {
     global $hyper_invalidated_post_id, $hyper_invalidated;
     
     if ($hyper_invalidated) 
     {
-        hyper_log("Already invalidated");
+        //hyper_log("Already invalidated");
         return;
     }
     if ($hyper_invalidated_post_id == $post_id) 
     {
-        hyper_log("Already invalidated post id $post_id");  
+        //hyper_log("Already invalidated post id $post_id");  
         return;
     }
     
     $post = get_post($post_id);
-    hyper_log("Post status " . $post->post_status);
+    //hyper_log("Post status " . $post->post_status);
     if ($post->post_status != 'publish') 
     {
         return;
     }
     $hyper_invalidated_post_id = $post_id;
     
-    
+    // Post invalidation
     $link = get_permalink($post_id);
     //$link = substr($link, strpos($link, '/', 7));
     $link = substr($link, 7);
     $file = md5($link);
 
-    if (file_exists(ABSPATH . 'wp-content/hyper-cache/' . $file . '.dat'))
-    {
-        unlink(ABSPATH . 'wp-content/hyper-cache/' . $file . '.dat');
-    }
-    if (file_exists(ABSPATH . 'wp-content/hyper-cache/pda' . $file . '.dat'))
-    {
-        unlink(ABSPATH . 'wp-content/hyper-cache/pda' . $file . '.dat');
-    }
-    if (file_exists(ABSPATH . 'wp-content/hyper-cache/iphone' . $file . '.dat'))
-    {
-        unlink(ABSPATH . 'wp-content/hyper-cache/iphone' . $file . '.dat');
-    }
+    unlink(ABSPATH . 'wp-content/hyper-cache/' . $file . '.dat');
+    unlink(ABSPATH . 'wp-content/hyper-cache/pda' . $file . '.dat');
+    unlink(ABSPATH . 'wp-content/hyper-cache/iphone' . $file . '.dat');
     
     // Home invalidation
     $link = substr(get_option('home'), 7) . '/';
     $file = md5($link);
 
-    @unlink(ABSPATH . 'wp-content/hyper-cache/' . $file . '.dat');
-    @unlink(ABSPATH . 'wp-content/hyper-cache/pda' . $file . '.dat');
-    @unlink(ABSPATH . 'wp-content/hyper-cache/iphone' . $file . '.dat');
-        
+    unlink(ABSPATH . 'wp-content/hyper-cache/' . $file . '.dat');
+    unlink(ABSPATH . 'wp-content/hyper-cache/pda' . $file . '.dat');
+    unlink(ABSPATH . 'wp-content/hyper-cache/iphone' . $file . '.dat');
 }
 
-function hyper_delete_path( $path = '' ) {
-	if ($handle = opendir($path)) {
-		while ($file = readdir($handle)) {
-			if ($file != '.' && $file != '..') {
+// Completely remove a directory and it's content.
+function hyper_delete_path($path) 
+{
+    if ($path == null) return;
+    
+	if ($handle = opendir($path)) 
+    {
+		while ($file = readdir($handle)) 
+        {
+			if ($file != '.' && $file != '..') 
+            {
 				unlink($path . '/' . $file);
 			}
 		}
@@ -251,12 +238,18 @@ function hyper_delete_path( $path = '' ) {
 	}
 }
 
-function hyper_count() {
+// Counts the number of file in to the hyper cache directory to give an idea of
+// the number of pages cached.
+function hyper_count() 
+{
     $count = 0;
-    if (!is_dir(ABSPATH . 'wp-content/hyper-cache')) return 0;
-	if ($handle = opendir(ABSPATH . 'wp-content/hyper-cache')) {
-		while ($file = readdir($handle)) {
-			if ($file != '.' && $file != '..') {
+    //if (!is_dir(ABSPATH . 'wp-content/hyper-cache')) return 0;
+	if ($handle = opendir(ABSPATH . 'wp-content/hyper-cache')) 
+    {
+		while ($file = readdir($handle)) 
+        {
+			if ($file != '.' && $file != '..') 
+            {
 				$count++;
 			}
 		}
@@ -265,18 +258,23 @@ function hyper_count() {
     return $count;
 }
 
-// Intercepts the action that can trigger a cache invalidation
+// Intercepts the action that can trigger a cache invalidation if the cache system is enabled
+// and invalidation is asked for actions (if is not only based on cache timeout)
 if ($hyper_options['enabled'] && $hyper_options['expire_type'] != 'none')
 {
-
     
-    // We need to invalidate everything for those actions: home page, categories pages, tags pages are affected
-	//add_action('publish_post', 'hyper_cache_invalidate_publish_post', 0);
-	//add_action('publish_phone', 'hyper_cache_invalidate', 0);
+    // We need to invalidate everything for those actions because home page, categories pages, 
+    // tags pages are affected and generally if we use plugin that print out "latest posts"
+    // or so we cannot know if a deleted post appears on every page.
     add_action('switch_theme', 'hyper_cache_invalidate', 0);
     add_action('delete_post', 'hyper_cache_invalidate', 0);
     
-	// Posts
+	// When a post is modified and we want to expire only it's page we listen for
+    // post edit (called everytime a post is modified, even if a comment is added) and
+    // the status change. We invalidate the single post if it's status is publish, we 
+    // invalidate all the cache if the status change from publish to "not publish" or
+    // from "not publish" to publish. These two cases make a post to appear or disappear
+    // anc can affect home, categories, single pages with a posts list, ...
     if ($hyper_options['expire_type'] == 'post')
     {
         add_action('edit_post', 'hyper_cache_invalidate_post', 0);
@@ -284,19 +282,22 @@ if ($hyper_options['enabled'] && $hyper_options['expire_type'] != 'none')
     }
     else
     {    
+        // If a complete invalidation is required, we do it on post edit.
         add_action('edit_post', 'hyper_cache_invalidate', 0);
     }
 
-    // Coment ID is received
-    //add_action('trackback_post', 'hyper_cache_invalidate', 0);
-    //add_action('pingback_post', 'hyper_cache_invalidate', 0);
+    // When a comment is received, and it's status is approved, the reference
+    // post is modified, but even other pages can be affected (last comments list,
+    // comment count and so on). We don't care about the latter situation when
+    // the expire type is configured to invalidate the single post page.
+    // Surely some of those hooks are redundant. When a new comment is added (and
+    // approved) the action "edit_post" is fired (llok at the code before). I need
+    // to deeply check this code BUT the plugin is protected from redundant invalidations.
     if ($hyper_options['expire_type'] == 'post')
     {    
         add_action('comment_post', 'hyper_cache_invalidate_comment', 10, 2);
         add_action('edit_comment', 'hyper_cache_invalidate_comment', 0);
         add_action('wp_set_comment_status', 'hyper_cache_invalidate_comment', 0);
-        
-        // No post_id is available
         add_action('delete_comment', 'hyper_cache_invalidate_comment', 0);
     }
     else 
@@ -304,12 +305,15 @@ if ($hyper_options['enabled'] && $hyper_options['expire_type'] != 'none')
         add_action('comment_post', 'hyper_cache_invalidate', 0);
         add_action('edit_comment', 'hyper_cache_invalidate', 0);
         add_action('wp_set_comment_status', 'hyper_cache_invalidate', 0);
-        
-        // No post_id is available
         add_action('delete_comment', 'hyper_cache_invalidate', 0);
     }
 }
 
+
+// Capture and register if a redirect is sent back from WP, so the cache
+// can cache (or ignore) it. Redirects were source of problems for blogs
+// with more than one host name (eg. domain.com and www.domain.com) comined
+// with the use of Hyper Cache.
 add_filter('redirect_canonical', 'hyper_redirect_canonical', 10, 2);
 $hyper_redirect = null;
 function hyper_redirect_canonical($redirect_url, $requested_url)
